# docs/swagger.yaml
...
definitions:
...
  config.ConfigReq:
    properties:
      clientUUID:
        type: string
      enableInternetAccess:
        type: boolean
      platformApiBase:
        description: Platform api url setting, e.g. "https://ao.space"`
        type: string
    type: object
  config.ConfigRsp:
    properties:
      connectedNetwork:
        items:
          $ref: '#/definitions/pair.Network'
        type: array
      enableInternetAccess:
        type: boolean
      enableLAN:
        type: boolean
      enableP2P:
        type: boolean
      userDomain:
        type: string
    type: object
  config.GetConfigRsp:
    properties:
      connectedNetwork:
        items:
          $ref: '#/definitions/pair.Network'
        type: array
      enableInternetAccess:
        type: boolean
      enableLAN:
        type: boolean
      enableP2P:
        type: boolean
      platformApiBase:
        type: string
      userDomain:
        type: string
    type: object
  create.CreateReq:
    properties:
      clientPhoneModel:
        description: 客户端手机型号.
        type: string
      clientUuid:
        description: 客户端 uuid
        type: string
      enableInternetAccess:
        description: 是否启用互联网通道
        type: boolean
      password:
        description: 空间密码
        type: string
      platformApiBase:
        description: Platform api url setting, e.g. "https://ao.space"`
        type: string
      spaceName:
        description: 空间名称
        type: string
      verificationMethod:
        description: did 的验证方法
        items:
          $ref: '#/definitions/document.VerificationMethod'
        type: array
    required:
    - clientPhoneModel
    - clientUuid
    - password
    - spaceName
    - verificationMethod
    type: object
  create.CreateRsp:
    properties:
      agentToken:
        description: agent 接口的访问 token. 暂未启用.
        type: string
      connectedNetwork:
        description: 设备连接的网络情况
        items:
          $ref: '#/definitions/pair.Network'
        type: array
      didDoc:
        description: did 文档
        type: string
      enableInternetAccess:
        description: 是否启用互联网通道
        type: boolean
      encryptedPriKeyBytes:
        description: 加密的空间密码凭证对应的私钥密文
        type: string
      spaceUserInfo:
        description: 空间信息
    type: object
...
  document.VerificationMethod:
    properties:
      id:
        type: string
      publicKeyPem:
        type: string
      type:
        type: string
    required:
    - id
    - publicKeyPem
    - type
    type: object
  dto.BaseRspStr:
    properties:
      code:
        type: string
      message:
        type: string
      requestId:
        type: string
      results: {}
    type: object
...
  network.NetworkAdapter:
    properties:
      adapterName:
        description: '网卡名称 eth0, eth1, wlan0 (获取网络信息: 返回;  其他: 不传;)'
        type: string
      connected:
        description: '是否已连接 (获取网络信息: 返回;  其他: 不传;)'
        type: boolean
      defaultGateway:
        description: '默认网关 (获取网络信息: 返回;  无线连上修改配置: 必传;  无线未连上修改配置:  可选;  有线时修改配置:  必传;)'
        type: string
      ipv4:
        description: 'ipv4 地址 (获取网络信息: 返回;  无线连上修改配置: 必传;  无线未连上修改配置:  可选;  有线时修改配置:  必传;)'
        type: string
      ipv4DNS1:
        description: 'ipv4 dNS1 地址 (获取网络信息: 返回; 其他: 选传)'
        type: string
      ipv4DNS2:
        description: 'ipv4 dNS2 地址 (获取网络信息: 返回; 其他: 选传)'
        type: string
      ipv4UseDhcp:
        description: 'ipv4 使用 dhcp 自动获取。 (获取网络信息: 返回;  其他: 必传;)'
        type: boolean
      ipv6:
        description: 'ipv6  (获取网络信息: 返回;  无线连上修改配置: 必传;  无线未连上修改配置:  可选;  有线时修改配置:  必传;)'
        type: string
      ipv6DefaultGateway:
        description: 'ipv6 默认网关 (获取网络信息: 返回;  无线连上修改配置: 必传;  无线未连上修改配置:  可选;  有线时修改配置:  必传;)'
        type: string
      ipv6UseDhcp:
        description: 'ipv6 使用 dhcp 自动获取 (获取网络信息: 返回;  其他: 必传;)'
        type: boolean
      mACAddress:
        description: '盒子网卡地址(不是路由器的网络地址)  (获取网络信息: 返回;  其他: 不传;)'
        type: string
      subNetMask:
        description: '子网掩码 (获取网络信息: 返回;  无线连上修改配置: 必传;  无线未连上修改配置:  可选;  有线时修改配置:  必传;)'
        type: string
      subNetPreLen:
        description: '子网前缀长度 (获取网络信息: 返回;  无线连上修改配置: 必传;  无线未连上修改配置:  可选;  有线时修改配置:  必传;)'
        type: string
      wIFIAddress:
        description: '路由器无线网络地址(不是盒子网卡地址)。有线连接时为空串。 (获取网络信息: 不返回;  无线连上修改配置: 不传;  无线未连上修改配置:  必传;  有线时修改配置:  不传;)'
        type: string
      wIFIName:
        description: 'WIFI名称。有线连接时为空串。 (获取网络信息: 返回;  无线连上修改配置: 必传;  无线未连上修改配置:  必传;  有线时修改配置:  不传;)'
        type: string
      wIFIPassword:
        description: 'WIFI密码。有线连接时为空串。 (获取网络信息: 不返回;  无线连上修改配置: 不传;  无线未连上修改配置:  必传;  有线时修改配置:  不传;)'
        type: string
      wired:
        description: '有线还是无线网卡。 true: 有线; false: 无线。 (获取网络信息: 返回;  其他: 必传;)'
        type: boolean
    type: object
  network.NetworkConfigReq:
    properties:
      dNS1:
        description: ipv4 dNS1 地址
        type: string
      dNS2:
        description: ipv4 dNS2 地址
        type: string
      ipv6DNS1:
        description: ipv6 dNS1 地址
        type: string
      ipv6DNS2:
        description: ipv6 dNS2 地址
        type: string
      networkAdapters:
        description: 网络适配器列表
        items:
          $ref: '#/definitions/network.NetworkAdapter'
        type: array
    type: object
  network.NetworkIgnoreReq:
    properties:
      wIFIName:
        description: WIFI名称。有线连接时为空串。
        type: string
    type: object
  network.NetworkStatusRsp:
    properties:
      dNS1:
        description: 'ipv4 dNS1 地址 (获取网络信息: 返回; 其他: 选传)'
        type: string
      dNS2:
        description: 'ipv4 dNS2 地址 (获取网络信息: 返回; 其他: 选传)'
        type: string
      internetAccess:
        description: '是否可以访问互联网 (获取网络信息: 返回;  其他: 不传;)'
        type: boolean
      ipv6DNS1:
        description: 'ipv6 dNS1 地址 (获取网络信息: 返回; 其他: 选传)'
        type: string
      ipv6DNS2:
        description: 'ipv6 dNS2 地址 (获取网络信息: 返回; 其他: 选传)'
        type: string
      lanAccessPort:
        description: '局域网访问网关服务的端口号 (获取网络信息: 返回;  其他: 不传;)'
        type: integer
      networkAdapters:
        description: '网络适配器列表 (获取网络信息: 返回; 其他: 必传)'
        items:
          $ref: '#/definitions/network.NetworkAdapter'
        type: array
    type: object
...
  pair.Network:
    properties:
      ip:
        type: string
      port:
        type: integer
      tlsPort:
        type: integer
      wifiName:
        type: string
      wire:
        description: 有线
        type: boolean
    type: object
...
  revoke.RevokeReq:
    properties:
      clientUuid:
        type: string
      password:
        type: string
    type: object
  revoke.RevokeRsp:
    properties:
      agentToken:
        type: string
      code:
        type: string
      message:
        type: string
      requestId:
        type: string
      results: {}
    type: object
...
info:
  contact: {}
paths:
...
  /agent/v1/api/bind/internet/service/config:
    get:
      consumes:
      - text/plain
      operationId: InternetServiceGetConfig
      parameters:
      - description: clientUuid and aoid
        in: query
        name: body
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code=AG-200 success;
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseRspStr'
            - properties:
                results:
                  $ref: '#/definitions/config.GetConfigRsp'
              type: object
      summary: get internet tunnel config [for client/gateway bluetooth/LAN]
      tags:
      - Pair
    post:
      consumes:
      - text/plain
      description: |-
        When the Internet tunnel is configured,
        system-agent starts to register the server with the platform, and when it succeeds,
        it calls the gateway interface, and the gateway registers the member information with the platform. And return the result to client.
      operationId: InternetServicePostConfig
      parameters:
      - description: config params
        in: body
        name: configReq
        required: true
        schema:
          $ref: '#/definitions/config.ConfigReq'
      produces:
      - application/json
      responses:
        "200":
          description: code=AG-200 success;
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseRspStr'
            - properties:
                results:
                  $ref: '#/definitions/config.ConfigRsp'
              type: object
      summary: config internet tunnel [for client/gateway bluetooth/LAN]
      tags:
      - Pair
...
  /agent/v1/api/bind/revoke:
    post:
      consumes:
      - text/plain
      operationId: BindRevoke
      parameters:
      - description: params
        in: body
        name: verifyReq
        required: true
        schema:
          $ref: '#/definitions/revoke.RevokeReq'
      produces:
      - application/json
      responses:
        "200":
          description: code=AG-200 success;
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseRspStr'
            - properties:
                results:
                  $ref: '#/definitions/revoke.RevokeRsp'
              type: object
      summary: admin unbind with server [for client bluetooth/LAN]
      tags:
      - Pair
  /agent/v1/api/bind/space/create:
    post:
      consumes:
      - text/plain
      description: |-
        This interface will submit the form data for space configuration on client,
        such as space identification information, administrator space password and server Internet service configuration, etc.
        When client calls this interface, it will synchronize the binding success status with server and store the status data of both sides. server will return the access_token to client.
        解绑后重新绑定情况下，创建空间时不传 SpaceName 和 EnableInternetAccess，传 Password 参数。
      operationId: SpaceCreate
      parameters:
      - description: admin password
        in: body
        name: createReq
        required: true
        schema:
          $ref: '#/definitions/create.CreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: code=AG-200 success; AG-577 register failed on self-host platform;
            AG-560 failed to create member
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseRspStr'
            - properties:
                results:
                  $ref: '#/definitions/create.CreateRsp'
              type: object
      summary: create aospace account(finished bind progress) [for client bluetooth/LAN]
      tags:
      - Pair
...
  /agent/v1/api/network/config:
    get:
      operationId: GETNetworkStatus
      produces:
      - application/json
      responses:
        "200":
          description: code=AG-200 success.
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseRspStr'
            - properties:
                results:
                  $ref: '#/definitions/network.NetworkStatusRsp'
              type: object
      summary: get network info。[for client LAN/call]
      tags:
      - network
    post:
      consumes:
      - application/json
      operationId: POSTNetworkConfig
      parameters:
      - description: 网络配置参数
        in: body
        name: networkConfigReq
        required: true
        schema:
          $ref: '#/definitions/network.NetworkConfigReq'
      produces:
      - application/json
      responses:
        "200":
          description: code=AG-200 成功.
          schema:
            $ref: '#/definitions/dto.BaseRspStr'
      summary: network config[for client LAN/Call]
      tags:
      - network
  /agent/v1/api/network/ignore:
    post:
      operationId: NetworkIgnore
      parameters:
      - description: params
        in: body
        name: networkIgnoreReq
        required: true
        schema:
          $ref: '#/definitions/network.NetworkIgnoreReq'
      produces:
      - application/json
      responses:
        "200":
          description: code=AG-200 success.
          schema:
            $ref: '#/definitions/dto.BaseRspStr'
      summary: ignore this network(SSID) ,delete net link。[for client LAN/call]
      tags:
      - network
...
swagger: "2.0"